fix(sessions): flush session-state before GitHub per-session save
The main-side GITHUB_SESSION_CONFIG_UPDATE handler looks sessions up in
session-state.json, which was only written on graceful close. Fresh
sessions returned "not-found" because they had no on-disk record yet.

Extract buildSessionState into session-persistence.ts and flush via
session.save before invoking updateSessionConfig, so the handler always
sees the current renderer view.
