+++ src/renderer/components/session/SessionGitHubConfig.tsx
@@ -3,6 +3,7 @@ import { useGitHubStore } from '../../stores/githubStore'
 import { useSessionStore } from '../../stores/sessionStore'
 import { trackUsage } from '../../stores/tipsStore'
 import { parseRepoUrlClient } from './parseRepoUrlClient'
+import { buildSessionState } from '../../session-persistence'
 import type { SessionGitHubIntegration } from '../../../shared/github-types'
 
 interface Props {
@@ -70,6 +71,11 @@ export default function SessionGitHubConfig({ sessionId, cwd, initial }: Props) 
       autoDetected: false,
     }
     try {
+      // Flush current sessions to session-state.json before invoking the
+      // main-side handler. session.save is otherwise only called on graceful
+      // close, so freshly-opened sessions aren't on disk — and the handler
+      // looks them up there and returns "not-found".
+      await window.electronAPI.session.save(buildSessionState())
       const r = await window.electronAPI.github.updateSessionConfig(sessionId, patch)
       if (r.ok) {
         // Mirror the patch into the renderer session store so the GitHub
